requirements.txt
pyspark
psycopg2-binary
python-dotenv
pandas
requests
boto3
s3fs 